python_analytics/requirements.txt
flask
flask-cors
mysql-connector-python
pandas
